Move Alice and Bob db states in separate modules
